refactor: remove unused code + renaming stuff for consistency
