+++ src/parse/parse.ts
@@ -2,10 +2,10 @@ import nearly from "nearley";
 import type { ParsedCatalogEntry } from "./types";
 import { HRowType, HSectionType } from "../tokenize";
 import type { HRow, TextRow, TokenizedCatalogEntry } from "../tokenize";
-import type { Major2, Section } from "../graduate-types";
 import { writeFile } from "fs/promises";
 import { FileName } from "../classify";
 import grammar from "./grammar";
+import type { Major2, Section } from "@/major2";
 
 export const parseRows = (rows: HRow[]) => {
   const parser = new nearly.Parser(nearly.Grammar.fromCompiled(grammar));
