+++ src/parse/types.ts
@@ -1,5 +1,5 @@
+import type { Major2 } from "@/major2";
 import { CatalogEntryType } from "../classify";
-import { type Major2 } from "../graduate-types";
 
 export type ParsedCatalogEntry = {
   url: URL;
