+++ src/runtime/types.ts
@@ -1,22 +1,7 @@
-import type { Result } from "../graduate-types/common";
-
 /** Represents the label for a stage in the scraper pipeline */
-export enum StageLabel {
+export enum PhaseLabel {
+  ScrapeMajorLinks = "Scrape Major Links",
   Classify = "Classify",
-  Filter = "Filter",
   Tokenize = "Tokenize",
-  SaveComment = "Save Comment",
   Parse = "Parse",
-  Save = "Save",
 }
-
-/**
- * Represents a pipeline value, at some point in the scraper pipeline. Contains
- * a trace of the stages that have been run, as well as the resulting value
- * (either an error, or OK), and the URL the pipeline was run on.
- */
-export type Pipeline<T> = {
-  id: URL;
-  trace: StageLabel[];
-  result: Result<T, unknown[]>;
-};
